Move simplify tests to subfolder
