+++ builder/test/simplify/BUILD
@@ -1,0 +1,32 @@
+package(
+    default_applicable_licenses = ["//:license"],
+    default_visibility = ["//visibility:private"],
+)
+
+cc_test(
+    name = "simplify",
+    srcs = [
+        "expr_generator.h",
+        "rule_tester.h",
+        "simplify.cc", 
+    ] + glob(["*_rules.cc"]),
+    deps = [
+        "//base/test:util",
+        "//builder",
+        "//builder:simplify_rules",
+        "//runtime",
+        "@googletest//:gtest_main",
+    ],
+    size = "small",
+)
+
+cc_test(
+    name = "substitute",
+    srcs = ["substitute.cc"],
+    deps = [
+        "//builder",
+        "//runtime",
+        "@googletest//:gtest_main",
+    ],
+    size = "small",
+)
